+++ tests/publisher.cpp
@@ -8,15 +8,18 @@
 #define BOOST_TEST_MODULE zeq_publisher
 
 #include "broker.h"
+#include <zeq/detail/broker.h>
+#include <zeq/detail/constants.h>
+#include <zeq/detail/sender.h>
 
 #include <servus/servus.h>
 
 BOOST_AUTO_TEST_CASE(create_uri_publisher)
 {
-    zeq::Publisher publisher( test::buildPublisherURI( ));
+    const zeq::Publisher publisher( zeq::URI( "" ));
 
-    const servus::URI& uri = publisher.getURI();
-    const std::string expectedScheme( "create-uri-publisher" );
+    const zeq::URI& uri = publisher.getURI();
+    const std::string expectedScheme( "tcp" );
     const std::string baseScheme = uri.getScheme().substr( 0,
                                                       expectedScheme.length( ));
     BOOST_CHECK_EQUAL( baseScheme, expectedScheme );
@@ -28,13 +31,13 @@ BOOST_AUTO_TEST_CASE(create_invalid_uri_publisher)
 {
     // invalid URI, hostname only not allowed
     BOOST_CHECK_THROW(
-        zeq::Publisher publisher( test::buildURI( "localhost" )),
+        zeq::Publisher publisher( zeq::URI( "localhost" )),
         std::runtime_error );
 }
 
 BOOST_AUTO_TEST_CASE(publish)
 {
-    zeq::Publisher publisher( test::buildPublisherURI( ));
+    zeq::Publisher publisher( zeq::NULL_SESSION );
 #ifdef ZEQ_USE_ZEROBUF
     BOOST_CHECK( publisher.publish( test::EchoOut( )));
 #endif
@@ -44,9 +47,8 @@ BOOST_AUTO_TEST_CASE(publish)
 
 BOOST_AUTO_TEST_CASE(publish_update_uri)
 {
-    zeq::URI uri = test::buildPublisherURI();
-    zeq::Publisher publisher( uri );
-    uri = publisher.getURI();
+    zeq::Publisher publisher( zeq::NULL_SESSION );
+    const zeq::URI& uri = publisher.getURI();
     BOOST_CHECK_MESSAGE( uri.getPort() != 0, uri );
     BOOST_CHECK_MESSAGE( !uri.getHost().empty(), uri );
 #ifdef ZEQ_USE_ZEROBUF
@@ -58,7 +60,7 @@ BOOST_AUTO_TEST_CASE(publish_update_uri)
 
 BOOST_AUTO_TEST_CASE(publish_empty_event)
 {
-    zeq::Publisher publisher( test::buildPublisherURI( ));
+    zeq::Publisher publisher( zeq::NULL_SESSION );
     BOOST_CHECK( publisher.publish( zeq::Event( zeq::vocabulary::EVENT_EXIT )));
 }
 
@@ -67,17 +69,113 @@ BOOST_AUTO_TEST_CASE(multiple_publisher_on_same_host)
     if( !servus::Servus::isAvailable() || getenv("TRAVIS"))
         return;
 
-    const zeq::uint128_t uuid = servus::make_UUID();
-    const std::string scheme = "schema-" + std::to_string( uuid.high( )) +
-                               std::to_string( uuid.low( ));
-    const servus::URI uri( scheme + "://*:0" );
+    const zeq::Publisher publisher1;
+    const zeq::Publisher publisher2;
+    const zeq::Publisher publisher3;
 
-    const zeq::Publisher publisher1( uri );
-    const zeq::Publisher publisher2( uri );
-    const zeq::Publisher publisher3( uri );
-
-    servus::Servus service( std::string( "_" ) + scheme + "._tcp" );
+    servus::Servus service( PUBLISHER_SERVICE );
     const servus::Strings& instances =
             service.discover( servus::Servus::IF_ALL, 1000 );
     BOOST_CHECK_EQUAL( instances.size(), 3 );
+}
+
+BOOST_AUTO_TEST_CASE(zeroconf_record)
+{
+    if( !servus::Servus::isAvailable() || getenv("TRAVIS"))
+        return;
+
+    const zeq::Publisher publisher;
+
+    servus::Servus service( PUBLISHER_SERVICE );
+    const servus::Strings& instances =
+            service.discover( servus::Servus::IF_ALL, 1000 );
+    BOOST_REQUIRE_EQUAL( instances.size(), 1 );
+
+    const std::string& instance = instances[0];
+    BOOST_CHECK_EQUAL( instance, publisher.getAddress( ));
+    BOOST_CHECK_EQUAL( service.get( instance, KEY_APPLICATION ), "publisher" );
+    BOOST_CHECK_EQUAL( zeq::uint128_t( service.get( instance, KEY_INSTANCE )),
+                       zeq::detail::Sender::getUUID( ));
+    BOOST_CHECK_EQUAL( service.get( instance, KEY_SESSION ), getUserName( ));
+    BOOST_CHECK_EQUAL( service.get( instance, KEY_USER ), getUserName( ));
+}
+
+BOOST_AUTO_TEST_CASE(custom_session)
+{
+    if( !servus::Servus::isAvailable() || getenv("TRAVIS"))
+        return;
+
+    const zeq::Publisher publisher( test::buildUniqueSession( ));
+
+    servus::Servus service( PUBLISHER_SERVICE );
+    const servus::Strings& instances =
+            service.discover( servus::Servus::IF_ALL, 1000 );
+    BOOST_REQUIRE_EQUAL( instances.size(), 1 );
+
+    const std::string& instance = instances[0];
+    BOOST_CHECK_EQUAL( service.get( instance, KEY_SESSION ),
+                       publisher.getSession( ));
+}
+
+BOOST_AUTO_TEST_CASE(different_session_at_runtime)
+{
+    if( !servus::Servus::isAvailable() || getenv("TRAVIS"))
+        return;
+
+    setenv( "ZEROEQ_SESSION", "testsession", 1 );
+    const zeq::Publisher publisher;
+
+    servus::Servus service( PUBLISHER_SERVICE );
+    const servus::Strings& instances =
+            service.discover( servus::Servus::IF_ALL, 1000 );
+    BOOST_REQUIRE_EQUAL( instances.size(), 1 );
+
+    const std::string& instance = instances[0];
+    BOOST_CHECK_EQUAL( service.get( instance, KEY_SESSION ),
+                       "testsession" );
+    unsetenv( "ZEROEQ_SESSION" );
+}
+
+BOOST_AUTO_TEST_CASE(empty_session)
+{
+    BOOST_CHECK_THROW( const zeq::Publisher publisher( "" ),
+                       std::runtime_error );
+}
+
+BOOST_AUTO_TEST_CASE(empty_session_from_environment)
+{
+    setenv( "ZEROEQ_SESSION", "", 1 );
+
+    const zeq::Publisher publisher;
+    BOOST_CHECK_EQUAL( publisher.getSession(), getUserName( ));
+
+    unsetenv( "ZEROEQ_SESSION" );
+}
+
+BOOST_AUTO_TEST_CASE(fixed_uri_and_session)
+{
+    if( !servus::Servus::isAvailable() || getenv("TRAVIS"))
+        return;
+
+    const zeq::Publisher publisher( zeq::URI( "127.0.0.1"),
+                                    test::buildUniqueSession( ));
+    servus::Servus service( PUBLISHER_SERVICE );
+    const servus::Strings& instances =
+            service.discover( servus::Servus::IF_ALL, 1000 );
+    BOOST_REQUIRE_EQUAL( instances.size(), 1 );
+}
+
+BOOST_AUTO_TEST_CASE(legacy_ctor)
+{
+    if( !servus::Servus::isAvailable() || getenv("TRAVIS"))
+        return;
+
+    const zeq::Publisher publisher( servus::URI( "foo://" ));
+
+    servus::Servus service( PUBLISHER_SERVICE );
+    const servus::Strings& instances =
+            service.discover( servus::Servus::IF_ALL, 1000 );
+    BOOST_CHECK_EQUAL( instances.size(), 1 );
+    BOOST_CHECK_EQUAL( service.get( instances[0], KEY_SESSION ),
+                       getUserName( ));
 }