+++ zeq/connection/broker.h
@@ -83,9 +83,6 @@ public:
     /** Destroy this broker. */
     ZEQ_API ~Broker();
 
-    /** @return the port to which this broker is bound. */
-    ZEQ_API uint16_t getPort() const;
-
     ZEQ_API std::string getAddress() const; //!< @internal
 
 private: